Fix Grammar in en.md for Headings Solution
A small tweak should be made to the sentence "Everything is obviously." to make it "Everything is obvious.", so that it has correct grammar. Then, the line "There's nothing strange, isn't it" doesn't really make sense so I thought it should be "There's nothing strange, right?". In the next line I added some "the"s and an "of" so the sentence flows better and doesn't read like it is notes. The third line was completely fine and left alone.